refactor: 어노테이션 네이밍 변경
