+++ src/main/java/com/sign/controller/support/JWTWrapped.java
@@ -1,4 +1,4 @@
-package com.sign.controller;
+package com.sign.controller.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ProtectedRequest {
+public @interface JWTWrapped {
 }
